Normalize XCOM 2 extension
Apply normalization checklist: convert to TypeScript, replace queryPath
with queryArgs, remove Bluebird, convert images to WebP, use framework
shorthands for mergeMods/steamAppId/SteamAPPId. Fixed original bug
where WOTC passed a mod path instead of game ID to supportedTools().

LAZ-333
